\name{get_genes}
\alias{get_genes}

\title{
 get gene vector from NCI RAS Pathway
}

\description{
 get gene vector from NCI RAS Pathway
}

\usage{
  get_genes(genes=NULL);
}

\arguments{
   \item{genes}{genes is a vector. If it is missing, all genes of the pathway will return.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

