\name{get_node_layout}
\alias{get_node_layout}

\title{
 get gene layout
}

\description{
 get gene layout
}

\usage{
  get_node_layout(genes=NULL);
}

\arguments{
   \item{genes}{genes is a vector. If it is missing, all gene layout will return.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

