\name{get_relations}
\alias{get_relations}

\title{
 get gene relations from NCI RAS Pathway
}

\description{
 get gene relations from NCI RAS Pathway
}

\usage{
  get_relations(genes=NULL);
}

\arguments{
   \item{genes}{genes is a vector. If it is missing, all gene relations of the pathway will return.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

