\name{grid_image}
\alias{grid_image}

\title{
 grid_image
}

\description{
 grid_image
}

\usage{
  grid_image <- function (image.file=file.p, grid.s=10, col=col[7], xy.lab=T);
}

\arguments{
   \item{image.file}{image file that should be readable by EBImage}
   \item{grid.s}{grid space size}
   \item{col}{grid line color}
   \item{xy.lab}{TRUE or FALSE to label grid number}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

